Paket für GUI-Hilfsklassen und Paket für NeuronalesNetz-Klassen
