+++ Daten/Daten-Figuren.dat
