+++ Daten/Daten-Kurve.dat
