+++ Daten/Daten-RGB.dat
