+++ Daten/Daten-Ziffern.dat
