+++ Daten/Netz-Kurve.nn
