+++ Daten/Netz-RGB.nn
