+++ Daten/Netz-Ziffern.nn
