+++ GUILib/MNISTDecoder.java
@@ -1,3 +1,5 @@
+package GUILib;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
