+++ GUILib/PanelKurve.java
@@ -1,8 +1,13 @@
+package GUILib;
+
 import java.awt.*;
 import java.awt.event.*;
 import java.awt.image.BufferedImage;
 import javax.swing.*;
 import javax.swing.event.*; 
+
+import NeuronalesNetz.*;
+
 
 public class PanelKurve extends JPanel {
