+++ GUILib/PanelZiffern.java
@@ -1,9 +1,14 @@
+package GUILib;
+
 import java.awt.*;
 import java.awt.event.*;
 import java.awt.image.BufferedImage;
 import javax.swing.*;
 import javax.swing.event.*; 
 import java.awt.RenderingHints;
+
+import NeuronalesNetz.*;
+
 
 public class PanelZiffern extends JPanel {
