+++ GUILib/package.bluej
@@ -1,0 +1,87 @@
+#BlueJ package file
+dependency1.from=PanelKurve
+dependency1.to=Datensammlung
+dependency1.type=UsesDependency
+dependency2.from=PanelKurve
+dependency2.to=Datensatz
+dependency2.type=UsesDependency
+dependency3.from=PanelRGB
+dependency3.to=Datensammlung
+dependency3.type=UsesDependency
+dependency4.from=PanelRGB
+dependency4.to=Datensatz
+dependency4.type=UsesDependency
+dependency5.from=PanelZiffern
+dependency5.to=Datensatz
+dependency5.type=UsesDependency
+dependency6.from=PanelFigur
+dependency6.to=Datensatz
+dependency6.type=UsesDependency
+objectbench.height=107
+objectbench.width=1095
+package.divider.horizontal=0.6
+package.divider.vertical=0.848
+package.editor.height=629
+package.editor.width=968
+package.editor.x=100
+package.editor.y=100
+package.frame.height=841
+package.frame.width=1119
+package.numDependencies=6
+package.numTargets=7
+package.showExtends=true
+package.showUses=true
+readme.height=60
+readme.name=@README
+readme.width=49
+readme.x=10
+readme.y=10
+target1.height=70
+target1.name=PanelFigur
+target1.showInterface=false
+target1.type=ClassTarget
+target1.width=120
+target1.x=280
+target1.y=90
+target2.height=70
+target2.name=MNISTDecoder
+target2.showInterface=false
+target2.type=ClassTarget
+target2.width=130
+target2.x=460
+target2.y=60
+target3.height=70
+target3.name=Datensammlung
+target3.showInterface=false
+target3.type=ClassTarget
+target3.width=150
+target3.x=260
+target3.y=340
+target4.height=70
+target4.name=PanelKurve
+target4.showInterface=false
+target4.type=ClassTarget
+target4.width=120
+target4.x=550
+target4.y=280
+target5.height=70
+target5.name=PanelRGB
+target5.showInterface=false
+target5.type=ClassTarget
+target5.width=120
+target5.x=10
+target5.y=270
+target6.height=70
+target6.name=Datensatz
+target6.showInterface=false
+target6.type=ClassTarget
+target6.width=120
+target6.x=770
+target6.y=200
+target7.height=70
+target7.name=PanelZiffern
+target7.showInterface=false
+target7.type=ClassTarget
+target7.width=120
+target7.x=10
+target7.y=140
