+++ NeuronalesNetz/TangensHyperbolicus.java
@@ -1,41 +1,43 @@
+package NeuronalesNetz;
+
 /**
- * Die Sigmoidfunktion als Aktivierungsfunktion.
+ * Der Tangenshyperbolicus als Aktivierungsfunktion.
  * 
  * @author Dr. Oliver Heidbüchel
  * @version 2023-11-24
  * @author Daniel Garmann
  * @version 2024-01-20
  */
-public class Sigmoid extends Aktivierungsfunktion
+public class TangensHyperbolicus extends Aktivierungsfunktion
 {
     /**
-     * Die Sigmoidfunktion.
+     * Die Tangens hyperbolicus Funktion.
      * 
      * @param x eine Zahl
-     * @return der Funktionswert der Sigmoidfunktion an der Stelle x
+     * @return der Funktionswert des Tangens hyperbolicus an der Stelle x
      */
     public double phi(double x){
-        return 1 / (1 + Math.exp(-x));
+        return Math.tanh(x);
     }
 
     /**
-     * Die Ableitung der Sigmoidfunktion.
+     * Die Ableitung der Tangens hyperbolicus Funktion.
      *
      * @param x eine Zahl
-     * @return die Ableitung der Sigmoidfunktion an der Stelle x
+     * @return die Ableitung des Tangens hyperbolicus an der Stelle x
      */
     public double phiStrich(double x){
         double phiVonX = phi(x);
-        return phiVonX * (1 - phiVonX);
+        return 1 - phiVonX * phiVonX;
     }
-
-    /**
+    
+        /**
      * Die Methode liefert eine String-Representation der Funktion
      * @return eine String-Repraesentation der Funktion
      */ 
     @Override
     public String toString() {
-        return "Sigmoid";
+        return "TanHyp";
     }
-}
 
+}
