--- Quellcode/package.bluej
@@ -1,287 +1,0 @@
-#BlueJ package file
-dependency1.from=Aktivierungsfunktion
-dependency1.to=Sigmoid
-dependency1.type=UsesDependency
-dependency10.from=PanelRGB
-dependency10.to=List
-dependency10.type=UsesDependency
-dependency11.from=GuiRGB
-dependency11.to=Netz
-dependency11.type=UsesDependency
-dependency12.from=GuiRGB
-dependency12.to=PanelRGB
-dependency12.type=UsesDependency
-dependency13.from=GuiRGB
-dependency13.to=Aktivierungsfunktion
-dependency13.type=UsesDependency
-dependency14.from=GuiRGB
-dependency14.to=Datensammlung
-dependency14.type=UsesDependency
-dependency15.from=PanelKurve
-dependency15.to=Datensammlung
-dependency15.type=UsesDependency
-dependency16.from=PanelKurve
-dependency16.to=Netz
-dependency16.type=UsesDependency
-dependency17.from=PanelKurve
-dependency17.to=Datensatz
-dependency17.type=UsesDependency
-dependency18.from=PanelKurve
-dependency18.to=List
-dependency18.type=UsesDependency
-dependency19.from=GuiKurve
-dependency19.to=Netz
-dependency19.type=UsesDependency
-dependency2.from=Aktivierungsfunktion
-dependency2.to=Softsign
-dependency2.type=UsesDependency
-dependency20.from=GuiKurve
-dependency20.to=PanelKurve
-dependency20.type=UsesDependency
-dependency21.from=GuiKurve
-dependency21.to=Aktivierungsfunktion
-dependency21.type=UsesDependency
-dependency22.from=GuiKurve
-dependency22.to=Datensammlung
-dependency22.type=UsesDependency
-dependency23.from=PanelFigur
-dependency23.to=Datensatz
-dependency23.type=UsesDependency
-dependency24.from=GuiFigur
-dependency24.to=Netz
-dependency24.type=UsesDependency
-dependency25.from=GuiFigur
-dependency25.to=Datensammlung
-dependency25.type=UsesDependency
-dependency26.from=GuiFigur
-dependency26.to=PanelFigur
-dependency26.type=UsesDependency
-dependency27.from=GuiFigur
-dependency27.to=Aktivierungsfunktion
-dependency27.type=UsesDependency
-dependency28.from=GuiFigur
-dependency28.to=List
-dependency28.type=UsesDependency
-dependency29.from=GuiFigur
-dependency29.to=Datensatz
-dependency29.type=UsesDependency
-dependency3.from=Aktivierungsfunktion
-dependency3.to=TangensHyperbolicus
-dependency3.type=UsesDependency
-dependency30.from=MNISTDecoder
-dependency30.to=List
-dependency30.type=UsesDependency
-dependency31.from=MNISTDecoder
-dependency31.to=Datensatz
-dependency31.type=UsesDependency
-dependency32.from=PanelZiffern
-dependency32.to=Datensatz
-dependency32.type=UsesDependency
-dependency33.from=GuiZiffern
-dependency33.to=Netz
-dependency33.type=UsesDependency
-dependency34.from=GuiZiffern
-dependency34.to=Datensammlung
-dependency34.type=UsesDependency
-dependency35.from=GuiZiffern
-dependency35.to=PanelZiffern
-dependency35.type=UsesDependency
-dependency36.from=GuiZiffern
-dependency36.to=Aktivierungsfunktion
-dependency36.type=UsesDependency
-dependency37.from=GuiZiffern
-dependency37.to=List
-dependency37.type=UsesDependency
-dependency38.from=GuiZiffern
-dependency38.to=Datensatz
-dependency38.type=UsesDependency
-dependency39.from=GuiZiffern
-dependency39.to=MNISTDecoder
-dependency39.type=UsesDependency
-dependency4.from=Netz
-dependency4.to=Aktivierungsfunktion
-dependency4.type=UsesDependency
-dependency40.from=Datensammlung
-dependency40.to=List
-dependency40.type=UsesDependency
-dependency41.from=Datensammlung
-dependency41.to=Datensatz
-dependency41.type=UsesDependency
-dependency42.from=Schicht
-dependency42.to=Aktivierungsfunktion
-dependency42.type=UsesDependency
-dependency43.from=Schicht
-dependency43.to=Neuron
-dependency43.type=UsesDependency
-dependency5.from=Netz
-dependency5.to=Schicht
-dependency5.type=UsesDependency
-dependency6.from=Neuron
-dependency6.to=Aktivierungsfunktion
-dependency6.type=UsesDependency
-dependency7.from=PanelRGB
-dependency7.to=Datensammlung
-dependency7.type=UsesDependency
-dependency8.from=PanelRGB
-dependency8.to=Netz
-dependency8.type=UsesDependency
-dependency9.from=PanelRGB
-dependency9.to=Datensatz
-dependency9.type=UsesDependency
-editor.fx.0.height=768
-editor.fx.0.width=1144
-editor.fx.0.x=-8
-editor.fx.0.y=-8
-objectbench.height=114
-objectbench.width=658
-package.divider.horizontal=0.6001805054151624
-package.divider.vertical=0.8188622754491018
-package.editor.height=527
-package.editor.width=980
-package.editor.x=0
-package.editor.y=0
-package.frame.height=768
-package.frame.width=1144
-package.numDependencies=43
-package.numTargets=19
-package.showExtends=true
-package.showUses=true
-project.charset=UTF-8
-readme.height=60
-readme.name=@README
-readme.width=49
-readme.x=10
-readme.y=10
-target1.height=70
-target1.name=Schicht
-target1.showInterface=false
-target1.type=ClassTarget
-target1.width=120
-target1.x=620
-target1.y=130
-target10.height=50
-target10.name=Neuron
-target10.showInterface=false
-target10.type=ClassTarget
-target10.width=80
-target10.x=710
-target10.y=230
-target11.height=70
-target11.name=TangensHyperbolicus
-target11.showInterface=false
-target11.type=ClassTarget
-target11.width=160
-target11.x=800
-target11.y=320
-target12.height=50
-target12.name=PanelRGB
-target12.showInterface=false
-target12.type=ClassTarget
-target12.width=100
-target12.x=10
-target12.y=220
-target13.height=50
-target13.name=Netz
-target13.showInterface=false
-target13.type=ClassTarget
-target13.width=80
-target13.x=540
-target13.y=60
-target14.height=70
-target14.name=List
-target14.showInterface=false
-target14.type=ClassTarget
-target14.width=150
-target14.x=100
-target14.y=390
-target15.height=40
-target15.name=Datensatz
-target15.showInterface=false
-target15.type=ClassTarget
-target15.width=90
-target15.x=90
-target15.y=30
-target16.height=50
-target16.name=GuiRGB
-target16.showInterface=false
-target16.type=ClassTarget
-target16.width=90
-target16.x=70
-target16.y=150
-target17.height=70
-target17.name=Softsign
-target17.showInterface=false
-target17.type=ClassTarget
-target17.width=120
-target17.x=830
-target17.y=230
-target18.height=50
-target18.name=GuiZiffern
-target18.showInterface=false
-target18.type=ClassTarget
-target18.width=90
-target18.x=450
-target18.y=150
-target19.height=70
-target19.name=Sigmoid
-target19.showInterface=false
-target19.type=ClassTarget
-target19.width=120
-target19.x=830
-target19.y=150
-target2.height=50
-target2.name=GuiFigur
-target2.showInterface=false
-target2.type=ClassTarget
-target2.width=90
-target2.x=330
-target2.y=150
-target3.height=40
-target3.name=Datensammlung
-target3.showInterface=false
-target3.type=ClassTarget
-target3.width=130
-target3.x=240
-target3.y=30
-target4.height=50
-target4.name=PanelKurve
-target4.showInterface=false
-target4.type=ClassTarget
-target4.width=100
-target4.x=140
-target4.y=220
-target5.height=70
-target5.name=Aktivierungsfunktion
-target5.showInterface=false
-target5.type=AbstractTarget
-target5.width=160
-target5.x=780
-target5.y=450
-target6.height=50
-target6.name=GuiKurve
-target6.showInterface=false
-target6.type=ClassTarget
-target6.width=90
-target6.x=200
-target6.y=150
-target7.height=50
-target7.name=PanelZiffern
-target7.showInterface=false
-target7.type=ClassTarget
-target7.width=90
-target7.x=400
-target7.y=220
-target8.height=50
-target8.name=MNISTDecoder
-target8.showInterface=false
-target8.type=ClassTarget
-target8.width=100
-target8.x=400
-target8.y=280
-target9.height=50
-target9.name=PanelFigur
-target9.showInterface=false
-target9.type=ClassTarget
-target9.width=90
-target9.x=270
-target9.y=220
