+++ package.bluej
@@ -1,0 +1,78 @@
+#BlueJ package file
+editor.fx.0.height=1080
+editor.fx.0.width=1870
+editor.fx.0.x=50
+editor.fx.0.y=0
+objectbench.height=116
+objectbench.width=1396
+package.divider.horizontal=0.6001805054151624
+package.divider.vertical=0.8305785123966942
+package.editor.height=596
+package.editor.width=1269
+package.editor.x=214
+package.editor.y=0
+package.frame.height=817
+package.frame.width=1420
+package.numDependencies=0
+package.numTargets=8
+package.showExtends=true
+package.showUses=true
+project.charset=UTF-8
+readme.height=60
+readme.name=@README
+readme.width=49
+readme.x=10
+readme.y=10
+target1.height=82
+target1.name=GUILib
+target1.type=PackageTarget
+target1.width=120
+target1.x=90
+target1.y=30
+target2.height=50
+target2.name=GuiFigur
+target2.showInterface=false
+target2.type=ClassTarget
+target2.width=90
+target2.x=330
+target2.y=150
+target3.height=70
+target3.name=MNIST
+target3.showInterface=false
+target3.type=ClassTarget
+target3.width=120
+target3.x=830
+target3.y=130
+target4.height=82
+target4.name=NeuronalesNetz
+target4.type=PackageTarget
+target4.width=130
+target4.x=230
+target4.y=30
+target5.height=50
+target5.name=GuiRGB
+target5.showInterface=false
+target5.type=ClassTarget
+target5.width=90
+target5.x=70
+target5.y=150
+target6.height=50
+target6.name=GuiZiffern
+target6.showInterface=false
+target6.type=ClassTarget
+target6.width=100
+target6.x=450
+target6.y=150
+target7.height=50
+target7.name=GuiKurve
+target7.showInterface=false
+target7.type=ClassTarget
+target7.width=90
+target7.x=200
+target7.y=150
+target8.height=70
+target8.name=README.md
+target8.type=TextTarget
+target8.width=120
+target8.x=10
+target8.y=280
